mm: Stop kswapd early when nothing's waiting for it to free pages
Keeping kswapd running when all the failed allocations that invoked it
are satisfied incurs a high overhead due to unnecessary page eviction
and writeback, as well as spurious VM pressure events to various
registered shrinkers. When kswapd doesn't need to work to make an
allocation succeed anymore, stop it prematurely to save resources.

Change-Id: Ife8d9dc651531a9f47affe9864b4bc0a3103cdec
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Samuel Pascua <pascua.samuel.14@gmail.com>
Signed-off-by: //. <174224977+skye-tachyon@users.noreply.github.com>